molecule/default/requirements.txt
yamllint==1.29.0
ansible==7.1.0
ansible-lint==6.14.0
flake8==6.0.0
molecule[podman,lint]==4.0.4
pytest-testinfra==7.0.0
